+++ lib/models/chess_arrow.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class ChessArrow {
+  final String startSquare;
+  final String endSquare;
+  final Color? color;
+
+  ChessArrow({
+    required this.startSquare,
+    required this.endSquare,
+    Color? color,
+  }) : color = color ?? Colors.amber.withOpacity(0.5);
+
+  @override
+  bool operator ==(Object other) {
+    return other is ChessArrow &&
+        startSquare == other.startSquare &&
+        endSquare == other.endSquare &&
+        color == other.color;
+  }
+
+  @override
+  int get hashCode =>
+      startSquare.hashCode ^ endSquare.hashCode ^ color.hashCode;
+}